Snapshot is downloaded in file prior to loading it into the store

+++ packages/chain/statemanager/sm_snapshots/downloader_test.go
@@ -136,7 +136,7 @@ func TestDownloaderReadTooMuch(t *testing.T) {
 	)
 }
 
-func testDownloader(t *testing.T, writeFun func(io.Writer), readFun func(io.Reader), chunkSize ...int) {
+func testDownloader(t *testing.T, writeFun func(io.Writer), readFun func(io.Reader), chunkSize ...uint64) {
 	log := testlogger.NewLogger(t)
 	defer log.Sync()
 	defer cleanupAfterDownloaderTest(t)
